template/requirements.txt
simplejson
azure-cosmos